Bug #42725 listener: Fix integer overflow
The bug happens when the free space is in a certain range:
 printf "%'08x\n" $(($(stat -f -c %a*%S /var/lib/univention-directory-listener)))
 # 1c0.000.000
   ^^
   |+- if bit31 is set, the right-shift creates a negative value
   +-- bits32+ are stripped because of the cast to (long)

Use int64_t.

Test:
 /etc/init.d/univention-directory-listener stop
 cd /var/lib/univention-directory-listener
 mount -t tmpfs -o size=7G xxx /var/lib/univention-directory-listener
 printf '%x\n' $(stat -f -c %a /var/lib/univention-directory-listener)
 # 1c0.000.000
 cp -pr * /var/lib/univention-directory-listener/
 /etc/runit/univention-directory-listener/run
 umount /var/lib/univention-directory-listener
 /etc/init.d/univention-directory-listener start

Merge r73434

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.2/ucs-4.2-0@73435 1b283449-9f2f-0410-b571-8eb228e1a901

+++ management/univention-directory-listener/src/notifier.c
@@ -42,6 +42,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <time.h>
+#include <inttypes.h>
 #include <sys/time.h>
 #include <sys/statvfs.h>
 
@@ -83,7 +84,7 @@ static int connect_to_ldap(univention_ldap_parameters_t *lp)
 
 static void check_free_space()
 {
-	static int min_mib = -2;
+	static int64_t min_mib = -2;
 	const char *dirnames[] = {
 		cache_dir,
 		ldap_dir,
@@ -98,16 +99,16 @@ static void check_free_space()
 
 	for (dirname=dirnames; *dirname; dirname++) {
 		struct statvfs buf;
-		int free_mib;
+		int64_t free_mib;
 
 		if (statvfs(*dirname, &buf))
 			continue;
 
-		free_mib = ((long)buf.f_bavail * (long)buf.f_frsize) >> 20;
+		free_mib = ((int64_t)buf.f_bavail * (int64_t)buf.f_frsize) >> 20;
 		if (free_mib >= min_mib)
 			continue;
 
-		univention_debug(UV_DEBUG_LISTENER, UV_DEBUG_ERROR, "File system '%s' full: %d < %d", *dirname, free_mib, min_mib);
+		univention_debug(UV_DEBUG_LISTENER, UV_DEBUG_ERROR, "File system '%s' full: %"PRId64" < %"PRId64, *dirname, free_mib, min_mib);
 		abort();
 	}
 }
